convert to gridder and degridder components
